Refactor code into modules.
